cloudrun/requirements.txt
Flask>=2.0
gunicorn>=20.0.4
google-cloud-pubsub>=2.0.0
nba_api>=1.1.10
requests>=2.20.0
pandas>=1.3.0
